Changed types to decrease memory usage

+++ include/datetime/Component.h
@@ -1,6 +1,8 @@
 #ifndef DATETIME_COMPONENT_H
 #define DATETIME_COMPONENT_H
 
+#include <cstdint>
+
 /**
  * Component of 'Date' or 'Time'.
  */
@@ -9,14 +11,14 @@ struct Component
     /**
      * Value of this 'Component'.
      */
-    int value;
+    uint32_t value;
 
     /**
      * Creates a 'Component' whose 'value' matches 'value'.
      *
      * @param value value to have the newly created 'Component' set to.
      */
-    explicit Component(int value) :
+    explicit Component(uint64_t value) :
         value(value) {}
 };
